# [Kernel] Layernorm performance optimization