@mawong-amd 2024-03-28
> @mawong-amd Thanks for submitting the PR! This optimization seems to be necessary for MI300x GPUs.
> 
> Unfortunately, I didn't see noticeable e2e performance boost for A100 GPUs. Is this expected? Also, I'm a bit worried about whether the new kernels keep the semantics of the current kernels. Could you double check?

Hi, I managed to run a few performance tests on H100 and I've also observed that there's no speed up. I looked at the PTX and SASS assembly and it looks like NVCC is not fusing the loads/stores as expected (here's an example with FP16 vector width = 8):

`st.global.u16 [%rd9], %rs5;
st.global.u16 [%rd9+2], %rs6;
st.global.u16 [%rd9+4], %rs11;
st.global.u16 [%rd9+6], %rs12;
st.global.u16 [%rd9+8], %rs17;
st.global.u16 [%rd9+10], %rs18;
st.global.u16 [%rd9+12], %rs23;
st.global.u16 [%rd9+14], %rs24;`

The good news is that I've confirmed that we can get similar speed ups on H100 (and probably A100) if I use another way to hint to the compiler to fuse these memory ops. As an example, by fusing the above into 4x st.global.u32, on LL2-70B FP16, input = 2048, output = 128, and batch = 1, preliminary numbers I'm seeing are prefill kernel times of 51 us (from 84 us) and decode kernel times of 3 us (from 8 us).

For now I'll mark the PR as a WIP and over the weekend I'll work to see if there's a convenient way to coerce NVCC to be a bit more aggressive with its memory op vectorization, be it via compiler flags ("-extra-device-vectorization" doesn't seem to do the trick) or refactoring the struct. If not, we'll probably have to move forward with platform specific implementations of said vectorization (preferably avoided to avoid code bloat).

As for the changed semantics, I'll discuss it in the relevant review comment thread.

Thanks!