@mawong-amd 2024-03-28
> @mawong-amd Thanks for submitting the PR! This optimization seems to be necessary for MI300x GPUs.
> 
> Unfortunately, I didn't see noticeable e2e performance boost for A100 GPUs. Is this expected? Also, I'm a bit worried about whether the new kernels keep the semantics of the current kernels. Could you double check?

Hi, I managed to run a few performance tests on H100 and I've also observed that there's no speed up. I looked at the PTX and SASS assembly and it looks like NVCC was not fusing the loads/stores as expected. It appears NVCC needs to know these global memory ops are aligned on a 16 byte boundary to unlock the full 128-bit coalesced op; I've added this alignment requirement to the vector struct and now I'm observing similar speedups on H100.

On LL2-70B FP16, input = 2048, output = 128, and batch = 1, preliminary numbers I'm seeing on H100 are prefill kernel times of 51 us (from 84 us) and decode kernel times of 3 us (from 8 us).

One "drawback" of this change is we can now only enable optimizations when the hidden_size is a multiple of 8 and the tensor pointers are aligned on a 16 byte boundary. But these conditions should be met essentially all the time.

As for the changed semantics, I'll discuss it in the relevant review comment thread.

Thanks!