@mawong-amd 2024-03-30
Quick update on end-to-end runtime numbers. With the latest changes, I'm seeing observable improvements on H100.

Specifically, on the latency benchmark:
- (input_len = output_len = batch_size = 128): Improves to 4.261s from 4.362s. [2.3% improvement]
- (input_len = 2048, output_len = 128, batch_size = 1): Improves to 11.463s from 11.658s. [1.7% improvement]