@youkaichao 2024-04-01
I realized that this pr breaks cuda 11.8 support because of the usage of `__half2` etc.