@mawong-amd 2024-04-01
> I realized that this pr breaks cuda 11.8 support because of the usage of `__half2` etc.

I think we can hotfix in a define guard to enable these optimizations only when the cuda version is > 11.8. Let me prepare a diff that does that.