@mawong-amd 2024-04-01
@youkaichao The patch is attached.
[layernorm_cuda11.8check.patch](https://github.com/vllm-project/vllm/files/14827660/layernorm_cuda11.8check.patch)